+++ src/tightdb/objc/test/shared_group.m
@@ -49,7 +49,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     // Read only shared group
     TDBContext* fromDisk = [TDBContext contextWithPersistenceToFile:@"employees.tightdb" error:nil];
 
-    [fromDisk readWithBlock:^(TDBTransaction* group) {
+    [fromDisk readUsingBlock:^(TDBTransaction* group) {
             SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
             for (size_t i = 0; i < [diskTable rowCount]; i++) {
@@ -60,7 +60,7 @@ TIGHTDB_TABLE_2(SharedTable2,
         }];
 
 
-    [fromDisk writeWithBlock:^(TDBTransaction* group) {
+    [fromDisk writeUsingBlock:^(TDBTransaction* group) {
             SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
             for (size_t i = 0; i < 50; i++) {
@@ -70,7 +70,7 @@ TIGHTDB_TABLE_2(SharedTable2,
         } error:nil];
 
 
-    [fromDisk writeWithBlock:^(TDBTransaction* group) {
+    [fromDisk writeUsingBlock:^(TDBTransaction* group) {
             SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
             for (size_t i = 0; i < 50; i++) {
@@ -80,7 +80,7 @@ TIGHTDB_TABLE_2(SharedTable2,
         } error:nil];
 
 
-    [fromDisk writeWithBlock:^(TDBTransaction* group) {
+    [fromDisk writeUsingBlock:^(TDBTransaction* group) {
             SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
             for (size_t i = 0; i < 50; i++) {
@@ -92,7 +92,7 @@ TIGHTDB_TABLE_2(SharedTable2,
             return YES; // commit
         } error:nil];
 
-    [fromDisk readWithBlock:^(TDBTransaction* group) {
+    [fromDisk readUsingBlock:^(TDBTransaction* group) {
             SharedTable2* diskTable = [group tableWithName:@"employees" asTableClass:[SharedTable2 class]];
             NSLog(@"Disktable size: %zu", [diskTable rowCount]);
         
@@ -112,7 +112,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     
     TDBContext* fromDisk = [TDBContext contextWithPersistenceToFile:@"readonlyTest.tightdb" error:nil];
     
-    [fromDisk writeWithBlock:^(TDBTransaction *group) {
+    [fromDisk writeUsingBlock:^(TDBTransaction *group) {
         TDBTable *t = [group createTableWithName:@"table"];
         
         [t addColumnWithName:@"col0" type:TDBIntType];
@@ -122,7 +122,7 @@ TIGHTDB_TABLE_2(SharedTable2,
         
     } error:nil];
     
-    [fromDisk readWithBlock:^(TDBTransaction* group) {
+    [fromDisk readUsingBlock:^(TDBTransaction* group) {
         TDBTable *t = [group tableWithName:@"table"];
         
         STAssertThrows([t addRow:nil], @"Is in readTransaction");
@@ -154,23 +154,23 @@ TIGHTDB_TABLE_2(SharedTable2,
 
     TDBContext* ctx = [TDBContext contextWithPersistenceToFile:@"singleTest.tightdb" error:nil];
 
-    [ctx writeWithBlock:^(TDBTransaction *trx) {
+    [ctx writeUsingBlock:^(TDBTransaction *trx) {
         TDBTable *t = [trx createTableWithName:@"table"];
         [t addColumnWithName:@"col0" type:TDBIntType];
         [t addRow:@[@10]];
         return YES;
     } error:nil];
 
-    [ctx readTable:@"table" withBlock:^(TDBTable* table) {
+    [ctx readTable:@"table" usingBlock:^(TDBTable* table) {
         STAssertTrue([table rowCount] == 1, @"No rows have been removed");
     }];
 
-    [ctx writeTable:@"table" withBlock:^(TDBTable* table) {
+    [ctx writeTable:@"table" usingBlock:^(TDBTable* table) {
         [table addRow:@[@10]];
         return YES;
     } error:nil];
 
-    [ctx readTable:@"table" withBlock:^(TDBTable* table) {
+    [ctx readTable:@"table" usingBlock:^(TDBTable* table) {
         STAssertTrue([table rowCount] == 2, @"Rows were added");
     }];
 }
@@ -188,7 +188,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     
     STAssertFalse([sg hasChangedSinceLastTransaction], @"SharedGroup has not changed");
     
-    [sg writeWithBlock:^(TDBTransaction* group) {
+    [sg writeUsingBlock:^(TDBTransaction* group) {
         [group createTableWithName:@"t"];
         return YES;
     } error:nil];
@@ -196,7 +196,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     STAssertFalse([sg hasChangedSinceLastTransaction], @"SharedGroup has not been changed by another process");
 
     
-    [sg writeWithBlock:^(TDBTransaction* group) {
+    [sg writeUsingBlock:^(TDBTransaction* group) {
         TDBTable *t = [group tableWithName:@"t"];
         [t addColumnWithName:@"col" type:TDBBoolType];
         [t addRow:nil];
@@ -212,7 +212,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     TDBContext *sg2 = [TDBContext contextWithPersistenceToFile:@"hasChanged.tightdb" error:nil];
     
     
-    [sg2 writeWithBlock:^(TDBTransaction* group) {
+    [sg2 writeUsingBlock:^(TDBTransaction* group) {
         TDBTable *t = [group tableWithName:@"t"];
         [t addRow:nil]; /* Adding an empty row */
         return YES;
@@ -231,7 +231,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     
     TDBContext *c = [TDBContext contextWithPersistenceToFile:contextPath error:nil];
     
-    [c writeWithBlock:^BOOL(TDBTransaction *transaction) {
+    [c writeUsingBlock:^BOOL(TDBTransaction *transaction) {
         
         STAssertThrows([transaction createTableWithName:nil], @"name is nil");
         STAssertThrows([transaction createTableWithName:@""], @"name is empty");
@@ -242,7 +242,7 @@ TIGHTDB_TABLE_2(SharedTable2,
         return YES;
     } error:nil];
     
-    [c readWithBlock:^(TDBTransaction *transaction) {
+    [c readUsingBlock:^(TDBTransaction *transaction) {
         
         STAssertThrows([transaction tableWithName:nil], @"name is nil");
         STAssertThrows([transaction tableWithName:@""], @"name is empty");
@@ -274,7 +274,7 @@ TIGHTDB_TABLE_2(SharedTable2,
         [context2 unpinReadTransactions];
     }
     {   // add something to the db to play with
-        [context1 writeWithBlock:^BOOL(TDBTransaction *transaction) {
+        [context1 writeUsingBlock:^BOOL(TDBTransaction *transaction) {
             TDBTable *t1 = [transaction createTableWithName:@"test"];
             [t1 addColumnWithName:@"col0" type:TDBBoolType];
             [t1 addRow:@[@YES]];
@@ -285,13 +285,13 @@ TIGHTDB_TABLE_2(SharedTable2,
     {   // validate that we can see previous commit from within a new pinned transaction
         BOOL changed = [context2 pinReadTransactions];
         STAssertTrue(changed, nil);
-        [context2 readWithBlock:^(TDBTransaction *transaction) {
+        [context2 readUsingBlock:^(TDBTransaction *transaction) {
             TDBTable *t = [transaction tableWithName:@"test"];
             STAssertEquals([[t rowAtIndex:0] boolInColumnWithIndex:0], YES, nil);
         }];
     }
     {   // commit new data in another context, without unpinning
-        [context1 writeWithBlock:^BOOL(TDBTransaction *transaction) {
+        [context1 writeUsingBlock:^BOOL(TDBTransaction *transaction) {
             TDBTable *t = [transaction tableWithName:@"test"];
             [t addRow:@[@NO]];
             return YES;
@@ -299,14 +299,14 @@ TIGHTDB_TABLE_2(SharedTable2,
         
     }
     {   // validate that we can see previous commit if we're not pinned
-        [context1 readWithBlock:^(TDBTransaction *transaction) {
+        [context1 readUsingBlock:^(TDBTransaction *transaction) {
             TDBTable *t = [transaction tableWithName:@"test"];
             STAssertEquals([[t rowAtIndex:1] boolInColumnWithIndex:0], NO, nil);
         }];
         
     }
      {   // validate that we can NOT see previous commit from within a pinned transaction
-        [context2 readWithBlock:^(TDBTransaction *transaction) {
+        [context2 readUsingBlock:^(TDBTransaction *transaction) {
             TDBTable *t = [transaction tableWithName:@"test"];
             STAssertEquals(t.rowCount, (NSUInteger)1, @"Still only 1 row");
         }];
@@ -316,7 +316,7 @@ TIGHTDB_TABLE_2(SharedTable2,
         [context2 unpinReadTransactions];
         BOOL changed = [context2 pinReadTransactions];
         STAssertTrue(changed, @"changes since last transaction");
-        [context2 readWithBlock:^(TDBTransaction *transaction) {
+        [context2 readUsingBlock:^(TDBTransaction *transaction) {
             TDBTable *t = [transaction tableWithName:@"test"];
             STAssertEquals(t.rowCount, (NSUInteger)2, @"Now we see 2 rows");
             STAssertEquals([[t rowAtIndex:1] boolInColumnWithIndex:0], NO, nil);
@@ -331,7 +331,7 @@ TIGHTDB_TABLE_2(SharedTable2,
 
     }
     {   // can't pin while we're inside a transaction
-        [context1 readWithBlock:^(TDBTransaction *transaction) {
+        [context1 readUsingBlock:^(TDBTransaction *transaction) {
             STAssertThrows([context1 pinReadTransactions], @"Can't pin inside transaction");
             STAssertNotNil(transaction, @"Parameter must be used");
         }];
@@ -339,7 +339,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     
     {   // can't unpin while we're inside a transaction
         [context1 pinReadTransactions];
-        [context1 readWithBlock:^(TDBTransaction *transaction) {
+        [context1 readUsingBlock:^(TDBTransaction *transaction) {
             STAssertThrows([context1 unpinReadTransactions], @"Can't unpin inside transaction");
             STAssertNotNil(transaction, @"Parameter must be used");
         }];
@@ -347,7 +347,7 @@ TIGHTDB_TABLE_2(SharedTable2,
     }
     {   // can't start a write transaction while pinned
         [context1 pinReadTransactions];
-        STAssertThrows([context1 writeWithBlock:^BOOL(TDBTransaction *transaction) {
+        STAssertThrows([context1 writeUsingBlock:^BOOL(TDBTransaction *transaction) {
             STAssertNotNil(transaction, @"Parameter must be used");
             return YES;
         } error:nil], @"Can't start write transaction while pinned");